put conditions all in one file
