--- lib/filter_builder/condition/base.rb
@@ -1,12 +1,0 @@
-module FilterBuilder
-  module Condition
-    class Base
-      attr_reader :field, :value
-
-      def initialize(field:, value:)
-        @field = field
-        @value = value
-      end
-    end
-  end
-end
