--- lib/filter_builder/condition/default.rb
@@ -1,9 +1,0 @@
-module FilterBuilder
-  module Condition
-    class Default < Base
-      def filter(scope)
-        scope.where(field.name => value)
-      end
-    end
-  end
-end
