--- lib/filter_builder/condition/does_not_equal.rb
@@ -1,9 +1,0 @@
-module FilterBuilder
-  module Condition
-    class DoesNotEqual < Base
-      def filter(scope)
-        scope.where.not(field.name => value)
-      end
-    end
-  end
-end
