--- lib/filter_builder/condition/does_not_match_case_sensitive.rb
@@ -1,9 +1,0 @@
-module FilterBuilder
-  module Condition
-    class DoesNotMatchCaseSensitive < Base
-      def filter(scope)
-        scope.where("#{field.namespaced} !~ ?", value)
-      end
-    end
-  end
-end
