--- lib/filter_builder/condition/equals.rb
@@ -1,9 +1,0 @@
-module FilterBuilder
-  module Condition
-    class Equals < Base
-      def filter(scope)
-        scope.where(field.name => value)
-      end
-    end
-  end
-end
