--- lib/filter_builder/condition/matches_case_insensitive.rb
@@ -1,9 +1,0 @@
-module FilterBuilder
-  module Condition
-    class MatchesCaseInsensitive < Base
-      def filter(scope)
-        scope.where("#{field.namespaced} ~* ?", value)
-      end
-    end
-  end
-end
